Finally fixed movements, added hitbox detection for player and mine

+++ src/main/java/Gravityhook/Abstract/GameObject.java
@@ -4,10 +4,10 @@ import Gravityhook.Interfaces.Drawable;
 import Gravityhook.Interfaces.IHitbox;
 
 public abstract class GameObject implements Drawable, IHitbox {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public GameObject(int x, int y) {
+    public GameObject(double x, double y) {
         this.x = x;
         this.y = y;
     }
